reset targets once they have stopped moving

+++ examples/toys/ping_pong_gun/createTargets.js
@@ -16,8 +16,9 @@ var scriptURL = Script.resolvePath('wallTarget.js');
 
 var MODEL_URL = 'http://hifi-public.s3.amazonaws.com/models/ping_pong_gun/target.fbx';
 var COLLISION_HULL_URL = 'http://hifi-public.s3.amazonaws.com/models/ping_pong_gun/target_collision_hull.obj';
+var MINIMUM_MOVE_LENGTH = 0.05;
+var RESET_DISTANCE = 0.5;
 
-var RESET_DISTANCE = 1;
 var TARGET_USER_DATA_KEY = 'hifi-ping_pong_target';
 var NUMBER_OF_TARGETS = 6;
 var TARGETS_PER_ROW = 3;
@@ -60,6 +61,8 @@ var targets = [];
 
 var originalPositions = [];
 
+var lastPositions = [];
+
 function addTargets() {
     var i;
     var row = -1;
@@ -77,6 +80,7 @@ function addTargets() {
         position.y = startPosition.y - (row * VERTICAL_SPACING);
 
         originalPositions.push(position);
+        lastPositions.push(position);
 
         var targetProperties = {
             name: 'Target',
@@ -103,7 +107,11 @@ function testTargetDistanceFromStart() {
         var distance = Vec3.subtract(originalPosition, currentPosition);
         var length = Vec3.length(distance);
 
-        if (length > RESET_DISTANCE) {
+        var moving = Vec3.length(Vec3.subtract(currentPosition, lastPositions[index]));
+
+        lastPositions[index] = currentPosition;
+
+        if (length > RESET_DISTANCE && moving<MINIMUM_MOVE_LENGTH) {
 
             Entities.deleteEntity(target);
 
@@ -121,6 +129,7 @@ function testTargetDistanceFromStart() {
             };
 
             targets[index] = Entities.addEntity(targetProperties);
+
         }
     });
 }
@@ -142,7 +151,7 @@ function deleteTargets() {
 }
 
 Entities.deletingEntity.connect(deleteEntity);
-var distanceCheckInterval = Script.setInterval(testTargetDistanceFromStart, 1000);
+var distanceCheckInterval = Script.setInterval(testTargetDistanceFromStart, 500);
 
 addTargets();
