+++ unpublishedScripts/masterReset.js
@@ -393,6 +393,7 @@ MasterReset = function() {
     var targets = [];
 
     var originalPositions = [];
+    var lastPositions = [];
 
     function addTargets() {
       var i;
@@ -441,7 +442,11 @@ MasterReset = function() {
         var distance = Vec3.subtract(originalPosition, currentPosition);
         var length = Vec3.length(distance);
 
-        if (length > RESET_DISTANCE) {
+        var moving = Vec3.length(Vec3.subtract(currentPosition, lastPositions[index]));
+
+        lastPositions[index] = currentPosition;
+
+        if (length > RESET_DISTANCE && moving < MINIMUM_MOVE_LENGTH) {
 
           Entities.deleteEntity(target);
 
@@ -455,15 +460,10 @@ MasterReset = function() {
             compoundShapeURL: COLLISION_HULL_URL,
             position: originalPositions[index],
             rotation: rotation,
-            script: targetsScriptURL,
-            userData: JSON.stringify({
-              resetMe: {
-                resetMe: true
-              }
-            })
+            script: scriptURL
           };
-          var target = Entities.addEntity(targetProperties);
-          targets[index] = target;
+
+          targets[index] = Entities.addEntity(targetProperties);
 
         }
       });
